Show if updates are available from AppLoader
