+++ src/locales/fr.json
@@ -7,6 +7,7 @@
 			"reset_data": "Réinitialisation des données...",
 			"config_keys_error": "Le fichier de configuration n'est pas le même que celui d'exemple. Certaines clés sont manquantes.",
 			"config_values_error": "Certaines clés du fichier de configuration ne sont pas renseignées. Cela conduira a des erreurs.",
+			"new_update": "Une nouvelle mise à jour est disponible (version %{version}). Cliquez ici pour l'installer",
 		},
 		"Home": {
 			"title": "Accueil",
@@ -24,7 +25,6 @@
 				"other": "dépensés cette semaine.",
 				"zero": "dépensé cette semaine."
 			},
-			"new_update": "Une nouvelle mise à jour est disponible (version %{version}). Cliquez ici pour l'installer",
 		},
 		"Refill": {
 			"title": "Rechargement",
