+++ src/screens/Home/HomeScreen.js
@@ -9,14 +9,13 @@
 import React from 'react';
 import { RefreshControl, Text, View } from 'react-native';
 import { connect } from 'react-redux';
-import VersionNumber from 'react-native-version-number';
 import List from '../../components/List';
 import colors from '../../styles/colors';
 import Balance from '../../components/Home/Balance';
 import Shortcuts from '../../components/Home/Shortcuts';
 import BlockTemplate from '../../components/BlockTemplate';
 import Item from '../../components/History/Item';
-import { GitHub, PayUTC } from '../../redux/actions';
+import { PayUTC } from '../../redux/actions';
 import { _, Home as t } from '../../utils/i18n';
 import { totalAmount } from '../../utils/stats';
 
@@ -38,14 +37,9 @@ class HomeScreen extends React.Component {
 	}
 
 	componentDidMount() {
-		const { dispatch, navigation } = this.props;
+		const { navigation } = this.props;
 
 		this.subscriptions = [navigation.addListener('willFocus', this.handleNavigationOnFocus)];
-
-		const action = GitHub.getLastestRelease();
-
-		dispatch(action);
-		action.payload.then(([release]) => this.checkNewUpdate(release));
 
 		this.onRefresh();
 	}
@@ -70,22 +64,6 @@ class HomeScreen extends React.Component {
 		this.setState({
 			message: params || {},
 		});
-	}
-
-	checkNewUpdate({ tag_name: tagName }) {
-		const { navigation } = this.props;
-		const { message } = this.state;
-		const { appVersion } = VersionNumber;
-
-		if (!message.message && appVersion && tagName && `v${appVersion}` !== tagName) {
-			this.setState({
-				message: {
-					message: t('new_update', { version: tagName }),
-					backgroundColor: colors.transfer,
-					onPress: () => navigation.navigate('Changelog'),
-				},
-			});
-		}
 	}
 
 	render() {
